Correcciones sobre certificado de etica
